Fix input validation visual bug
Fixed input validation indicator visual bug. It uses javascipt in order
to calculate the height of each <li> element. Then it multiplies them
by its parent's element height (offset height to be specific). This
value is stored until it is used to expand/shrink the reminder for the
validation UI.